@@ -54,21 +81,11 @@
     "aegir": "^36.0.2",
     "benchmark": "^2.1.4",
     "sinon": "^12.0.1",
-    "util": "^0.12.3"
-  },
-  "aegir": {
-    "build": {
-      "bundlesizeMax": "71kB"
-    }
+    "util": "^0.12.3",
+    "wherearewe": "^1.0.0"
   },
   "engines": {
-    "node": ">=12.0.0"
-  },
-  "eslintConfig": {
-    "extends": "ipfs",
-    "ignorePatterns": [
-      "src/*.d.ts"
-    ]
+    "node": ">=15.0.0"
   },
   "repository": {
     "type": "git",
